requirements.txt
bkcharts==0.2
bokeh==0.12.7
Django==1.11.4
django-picklefield==1.0.0
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
pandas==0.20.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
six==1.10.0
tornado==4.5.2
